requirements.txt
aiogram
loguru
python-dotenv
aiohttp